+++ array/static-vector-math.h
@@ -2,7 +2,7 @@
 
 DECLARE_NAMESPACE_NLL
 
-template <class Type, int N>
+template <class Type, size_t N>
 StaticVector<Type, N> sqrt(const StaticVector<Type, N>& v)
 {
    StaticVector<Type, N> r(no_init_tag); // make sure the vector is not initialized
@@ -13,7 +13,7 @@ StaticVector<Type, N> sqrt(const StaticVector<Type, N>& v)
    return r;
 }
 
-template <class Type, int N>
+template <class Type, size_t N>
 StaticVector<Type, N> sqr(const StaticVector<Type, N>& v)
 {
    StaticVector<Type, N> r(no_init_tag); // make sure the vector is not initialized
@@ -24,7 +24,7 @@ StaticVector<Type, N> sqr(const StaticVector<Type, N>& v)
    return r;
 }
 
-template <class Type, int N, class TypeSum = Type>
+template <class Type, size_t N, class TypeSum = Type>
 TypeSum sum(const StaticVector<Type, N>& v)
 {
    return std::accumulate(v.begin(), v.end(), TypeSum());
@@ -44,25 +44,25 @@ StaticVector<T, 3> cross(const StaticVector<T, 3>& a, const StaticVector<T, 3>& 
    return res;
 }
 
-template <class T, int N>
+template <class T, size_t N>
 T dot(const StaticVector<T, N>& lhs, const StaticVector<T, N>& rhs)
 {
    T accum = 0;
-   for (int n = 0; n < N; ++n)
+   for (size_t n = 0; n < N; ++n)
       accum += lhs[n] * rhs[n];
    return accum;
 }
 
-template <class T, int N>
+template <class T, size_t N>
 T norm2sqr(const StaticVector<T, N>& lhs)
 {
    T accum = 0;
-   for (int n = 0; n < N; ++n)
+   for (size_t n = 0; n < N; ++n)
       accum += lhs[n] * lhs[n];
    return accum;
 }
 
-template <class T, int N>
+template <class T, size_t N>
 T norm2(const StaticVector<T, N>& lhs)
 {
    return static_cast<T>(std::sqrt(norm2sqr(lhs)));
